requirements.txt
# Thumper Counter - Python Dependencies
# Core backend dependencies for deer tracking system

# Database
sqlalchemy>=2.0.0,<3.0.0          # ORM for database models
psycopg2-binary>=2.9.0            # PostgreSQL driver
alembic>=1.12.0                   # Database migrations

# Web Framework
fastapi>=0.104.0                  # Modern web framework
uvicorn[standard]>=0.24.0         # ASGI server
python-multipart>=0.0.6           # Form data support
pydantic>=2.0.0                   # Data validation

# Task Queue
celery>=5.3.0                     # Distributed task queue
redis>=5.0.0                      # Redis client for Celery broker
flower>=2.0.0                     # Celery monitoring

# Machine Learning
torch>=2.0.0                      # PyTorch for ML models
torchvision>=0.15.0               # Computer vision models
ultralytics>=8.0.0                # YOLOv8 for detection
pillow>=10.0.0                    # Image processing (needed for EXIF extraction)
opencv-python>=4.8.0              # OpenCV for image processing
numpy>=1.24.0                     # Numerical computing

# Utilities
python-dotenv>=1.0.0              # Environment variable loading
httpx>=0.25.0                     # HTTP client for async requests

# Development & Testing
pytest>=7.4.0                     # Testing framework
pytest-asyncio>=0.21.0            # Async test support
pytest-cov>=4.1.0                 # Test coverage

# Code Quality
black>=23.0.0                     # Code formatting
flake8>=6.0.0                     # Linting
mypy>=1.5.0                       # Type checking
Pillow==10.1.0
